make crc use unsigned short instead of unsigned int tidy up handling of crc type by using crc_t
git-svn-id: tumanako.svn.sourceforge.net::svn/tumanako/trunk@526 c3429617-7ab9-49b5-8233-5f90b2ce515f

+++ src/c/crc.h
@@ -38,7 +38,7 @@ extern "C" {
  *
  * This type must be big enough to contain at least 16 bits.
  *****************************************************************************/
-typedef unsigned int crc_t;
+typedef unsigned short crc_t;
 
 /**
  * Calculate the initial crc value.
